refactor(auth): rename AuthShell → Authenticator; wrap children
Authenticator now accepts children and always renders them. When the
session ends it keeps children mounted but makes them pointer-events-none
and select-none, then renders the logged-out modal via Radix Portal
(which escapes the disabled subtree and stays interactive).
