+++ web/src/app/layout.tsx
@@ -19,7 +19,7 @@ import StatsOverlayLoader from "@/components/dev/StatsOverlayLoader";
 import { cn } from "@opal/utils";
 import AppHealthBanner from "@/sections/banners/HealthBanner";
 import LicenseExpiryBanner from "@/sections/banners/LicenseExpiryBanner";
-import AuthShell from "@/sections/AuthShell";
+import Authenticator from "@/sections/Authenticator";
 import ProductGatingWrapper from "@/providers/ProductGatingWrapper";
 import SWRConfigProvider from "@/providers/SWRConfigProvider";
 
@@ -124,20 +124,21 @@ export default function Layout({ children }: LayoutProps) {
               <PHProvider>
                 <SWRConfigProvider>
                   <AppHealthBanner />
-                  <AuthShell />
                   <LicenseExpiryBanner />
-                  <AppProvider>
-                    <PostHogRuntimeInitializer />
-                    <CustomAnalyticsScript />
-                    <PostHogPageTracker />
-                    <div id={MODAL_ROOT_ID} className="h-screen w-screen">
-                      <ProductGatingWrapper>{children}</ProductGatingWrapper>
-                    </div>
-                    <WebVitals />
-                    {process.env.NEXT_PUBLIC_ENABLE_STATS === "true" && (
-                      <StatsOverlayLoader />
-                    )}
-                  </AppProvider>
+                  <Authenticator>
+                    <AppProvider>
+                      <PostHogRuntimeInitializer />
+                      <CustomAnalyticsScript />
+                      <PostHogPageTracker />
+                      <div id={MODAL_ROOT_ID} className="h-screen w-screen">
+                        <ProductGatingWrapper>{children}</ProductGatingWrapper>
+                      </div>
+                      <WebVitals />
+                      {process.env.NEXT_PUBLIC_ENABLE_STATS === "true" && (
+                        <StatsOverlayLoader />
+                      )}
+                    </AppProvider>
+                  </Authenticator>
                 </SWRConfigProvider>
               </PHProvider>
             </TooltipProvider>
