--- web/src/sections/AuthShell.tsx
@@ -1,44 +1,0 @@
-"use client";
-
-import { useRouter } from "next/navigation";
-import { useSessionWatcher } from "@/lib/auth/hooks";
-import { getExtensionContext } from "@/lib/extension/utils";
-import Modal from "@/refresh-components/Modal";
-import { Button, Text } from "@opal/components";
-import { SvgLogOut } from "@opal/icons";
-
-export default function AuthShell() {
-  const router = useRouter();
-  const sessionEnded = useSessionWatcher();
-
-  function handleLogin() {
-    const { isExtension } = getExtensionContext();
-    if (isExtension) {
-      window.open(
-        window.location.origin + "/auth/login",
-        "_blank",
-        "noopener,noreferrer"
-      );
-    } else {
-      router.push("/auth/login");
-    }
-  }
-
-  if (!sessionEnded) return null;
-
-  return (
-    <Modal open>
-      <Modal.Content width="sm" height="sm">
-        <Modal.Header icon={SvgLogOut} title="You Have Been Logged Out" />
-        <Modal.Body>
-          <Text font="main-ui-body" color="text-03">
-            Your session has expired. Please log in again to continue.
-          </Text>
-        </Modal.Body>
-        <Modal.Footer>
-          <Button onClick={handleLogin}>Log In</Button>
-        </Modal.Footer>
-      </Modal.Content>
-    </Modal>
-  );
-}
